requirements.txt
tensorflow==2.13.0
t-tech-investments
pandas
numpy
flask
redis
schedule
python-dotenv
